refactor: Split surrealdb service to new crate
