+++ core/core/Cargo.toml
@@ -87,7 +87,6 @@ services-redis-native-tls = ["services-redis", "redis?/tokio-native-tls-comp"]
 services-rocksdb = ["dep:rocksdb", "internal-tokio-rt"]
 services-seafile = []
 services-sftp = ["dep:openssh", "dep:openssh-sftp-client", "dep:fastpool"]
-services-surrealdb = ["dep:surrealdb"]
 services-swift = []
 services-upyun = ["dep:hmac", "dep:sha1"]
 services-vercel-artifacts = []
@@ -177,8 +176,6 @@ redis = { version = "1.0", features = [
 rocksdb = { version = "0.21.0", default-features = false, optional = true }
 # for services-hdfs-native
 hdfs-native = { version = "0.13", optional = true }
-# for services-surrealdb
-surrealdb = { version = "2", optional = true, features = ["protocol-http"] }
 # for services-monoiofs
 flume = { version = "0.11", optional = true }
 monoio = { version = "0.2.4", optional = true, features = [
