+++ core/core/src/services/mod.rs
@@ -129,11 +129,6 @@ mod sftp;
 #[cfg(feature = "services-sftp")]
 pub use sftp::*;
 
-#[cfg(feature = "services-surrealdb")]
-mod surrealdb;
-#[cfg(feature = "services-surrealdb")]
-pub use surrealdb::*;
-
 #[cfg(feature = "services-swift")]
 mod swift;
 #[cfg(feature = "services-swift")]
