+++ core/services/surrealdb/src/config.rs
@@ -61,10 +61,10 @@ impl Debug for SurrealdbConfig {
     }
 }
 
-impl crate::Configurator for SurrealdbConfig {
+impl opendal_core::Configurator for SurrealdbConfig {
     type Builder = SurrealdbBuilder;
 
-    fn from_uri(uri: &crate::types::OperatorUri) -> crate::Result<Self> {
+    fn from_uri(uri: &opendal_core::OperatorUri) -> opendal_core::Result<Self> {
         let mut map = uri.options().clone();
 
         if let Some(authority) = uri.authority() {
@@ -112,8 +112,8 @@ impl crate::Configurator for SurrealdbConfig {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Configurator;
-    use crate::types::OperatorUri;
+    use opendal_core::Configurator;
+    use opendal_core::OperatorUri;
 
     #[test]
     fn from_uri_sets_connection_namespace_database_table_and_root() {
