+++ core/src/lib.rs
@@ -90,6 +90,8 @@ pub mod services {
     pub use opendal_service_sled::*;
     #[cfg(feature = "services-sqlite")]
     pub use opendal_service_sqlite::*;
+    #[cfg(feature = "services-surrealdb")]
+    pub use opendal_service_surrealdb::*;
     #[cfg(feature = "services-tikv")]
     pub use opendal_service_tikv::*;
     #[cfg(feature = "services-vercel-blob")]
